CB-3282 save data sources through rm (#1742)
* CB-3282 save data sources through rm

* CB-3282 use rm storage only for distributed products

* CB-3282 fix npe for anonymous projects

* CB-3282 merge devel

---------
