+++ server/bundles/io.cloudbeaver.model/src/io/cloudbeaver/WebProjectImpl.java
@@ -16,33 +16,19 @@
  */
 package io.cloudbeaver;
 
-import io.cloudbeaver.model.rm.RMUtils;
-import org.eclipse.core.resources.IProject;
 import org.jkiss.code.NotNull;
-import org.jkiss.code.Nullable;
+import org.jkiss.dbeaver.DBException;
+import org.jkiss.dbeaver.Log;
 import org.jkiss.dbeaver.model.app.DBPDataSourceRegistry;
 import org.jkiss.dbeaver.model.app.DBPWorkspace;
 import org.jkiss.dbeaver.model.auth.SMSessionContext;
 import org.jkiss.dbeaver.model.rm.RMController;
 import org.jkiss.dbeaver.model.rm.RMProject;
-import org.jkiss.dbeaver.registry.BaseProjectImpl;
-import org.jkiss.dbeaver.registry.DataSourceRegistry;
-import org.jkiss.utils.CommonUtils;
-import org.jkiss.utils.Pair;
+import org.jkiss.dbeaver.registry.rm.DataSourceRegistryRM;
+import org.jkiss.dbeaver.runtime.DBWorkbench;
 
-import java.nio.file.Path;
-import java.util.Collection;
-
-public class WebProjectImpl extends BaseProjectImpl {
-
-    @NotNull
-    private final RMProject project;
-
-    @NotNull
-    private final Path path;
-    @NotNull
-    protected final DataSourceFilter dataSourceFilter;
-    private final RMController resourceController;
+public class WebProjectImpl extends BaseWebProjectImpl {
+    private static final Log log = Log.getLog(WebProjectImpl.class);
 
     public WebProjectImpl(
         @NotNull DBPWorkspace workspace,
@@ -51,114 +37,40 @@ public class WebProjectImpl extends BaseProjectImpl {
         @NotNull RMProject project,
         @NotNull DataSourceFilter dataSourceFilter
     ) {
-        super(workspace, sessionContext);
-        this.resourceController = resourceController;
-        this.path = RMUtils.getProjectPath(project);
-        this.project = project;
-        this.dataSourceFilter = dataSourceFilter;
-    }
-
-    @NotNull
-    public RMController getResourceController() {
-        return resourceController;
+        super(workspace, resourceController, sessionContext, project, dataSourceFilter);
     }
 
     @Override
-    public boolean isVirtual() {
-        return true;
-    }
-
-    @NotNull
-    @Override
-    public String getName() {
-        return project.getDisplayName();
-    }
-
-    @NotNull
-    @Override
-    public String getId() {
-        return project.getId();
-    }
-
-    @NotNull
-    @Override
-    public Path getAbsolutePath() {
-        return path;
-    }
-
-    @Nullable
-    @Override
-    public IProject getEclipseProject() {
-        return null;
+    public Object getProjectProperty(String propName) {
+        try {
+            return getResourceController().getProjectProperty(getId(), propName);
+        } catch (DBException e) {
+            log.error("Cannot get project property", e);
+            return null;
+        }
     }
 
     @Override
-    public boolean isOpen() {
-        return true;
-    }
-
-    @Override
-    public void ensureOpen() {
-
+    public void setProjectProperty(String propName, Object propValue) {
+        try {
+            getResourceController().setProjectProperty(getId(), propName, propValue);
+        } catch (DBException e) {
+            log.error("Cannot set project property", e);
+        }
     }
 
     @Override
     public boolean isUseSecretStorage() {
-        return false;
-    }
-
-    @NotNull
-    public RMProject getRmProject() {
-        return this.project;
+        return DBWorkbench.isDistributed();
     }
 
     @NotNull
     @Override
     protected DBPDataSourceRegistry createDataSourceRegistry() {
-        return new WebDataSourceRegistryProxy(new DataSourceRegistry(this), dataSourceFilter);
+        return new WebDataSourceRegistryProxy(
+            new DataSourceRegistryRM(this, getResourceController()),
+            dataSourceFilter
+        );
     }
 
-    /**
-     * Method for Bulk Update of resources properties paths
-     *
-     * @param oldToNewPaths collection of OldPath to NewPath pairs
-     */
-    public void moveResourcePropertiesBatch(@NotNull Collection<Pair<String, String>> oldToNewPaths) {
-        loadMetadata();
-        synchronized (metadataSync) {
-            for (var pathsPair : oldToNewPaths) {
-                final var oldResourcePath = CommonUtils.normalizeResourcePath(pathsPair.getFirst());
-                final var newResourcePath = CommonUtils.normalizeResourcePath(pathsPair.getSecond());
-                final var resProps = resourceProperties.remove(oldResourcePath);
-                if (resProps != null) {
-                    resourceProperties.put(newResourcePath, resProps);
-                }
-            }
-        }
-        flushMetadata();
-    }
-
-    /**
-     * Method for Bulk Remove of resources properties
-     */
-    public boolean resetResourcesPropertiesBatch(@NotNull Collection<String> resourcesPaths) {
-        loadMetadata();
-        boolean propertiesChanged = false;
-        synchronized (metadataSync) {
-            for (var resourcePath : resourcesPaths) {
-                var removedProperties = resourceProperties.remove(CommonUtils.normalizeResourcePath(resourcePath));
-                if (removedProperties != null) {
-                    propertiesChanged = true;
-                }
-            }
-        }
-        if (propertiesChanged) {
-            flushMetadata();
-        }
-        return propertiesChanged;
-    }
-
-    public Path getMetadataFilePath() {
-        return getMetadataPath().resolve(METADATA_STORAGE_FILE);
-    }
 }
